requirements.txt
apache-airflow-providers-telegram
apache-airflow-providers-smtp
apache-airflow-providers-postgres
apache-airflow-providers-amazon

apache-airflow-providers-jdbc
apache-airflow-providers-google
apache-airflow-providers-odbc
apache-airflow-providers-common-sql
apache-airflow-providers-mongo

pandas==2.0.3
beautifulsoup4==4.12.2
tqdm==4.60.0
xlrd==1.2.0
great-expectations==0.17.2
ipdb==0.13.13
inflect==7.0.0
boto3==1.34.19
Unidecode==1.3.8
Jinja2==3.1.3
weasyprint==57.1
pybase64==1.3.2
geopandas==0.14.3
matplotlib==3.8.3
plotly==5.19.0
Faker==23.3.0
kaleido==0.2.1
minio==7.2.5
pymongo==3.11.0  # ou qualquer versão específica que você precise
selenium==3.141.0  # Versão estável mais recente do Selenium
webdriver-manager==3.5.3 
webdriver_manager

